# control-api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Media Inference Engine
  description: 'Two surfaces on the same server:


    - **Engine `/v1`** — catalog operations, job queue, assets, text chat.

    - **Control `/api/v1`** — GPU profiles, catalog downloads, generate / upscale / LTX, cleanup.


    Mac proxy: `http://127.0.0.1:8090` → GPU `http://192.168.50.100:8090`.

    Interactive docs: `/docs` (Swagger) and `/redoc`.'
  version: 3.0.0
servers:
- url: http://127.0.0.1:8090
  description: Mac proxy (LAN)
- url: http://192.168.50.100:8090
  description: GPU box (direct)
paths:
...
  /api/v1/generate:
    post:
      tags:
      - Control UI (legacy /api/v1)
      summary: Queue image generation
      description: Multi-prompt batch (`prompts[]`) or shorthand (`prompt` + `count`). Gemma plans once,
        then Comfy runs items. Partial failures continue.
      operationId: api_generate_post_api_v1_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/generate/{job_id}/images/{filename}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Download generate output image
      operationId: api_generate_image_api_v1_generate__job_id__images__filename__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: filename
        in: path
        required: true
        schema:
          type: string
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflows/character-master:
    post:
      tags:
      - Control UI (legacy /api/v1)
      summary: Queue Character Master candidates
      operationId: api_character_master_post_api_v1_workflows_character_master_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterMasterRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflows/character-master/{job_id}/images/{filename}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Download Character Master image
      operationId: api_character_master_image_api_v1_workflows_character_master__job_id__images__filename__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: filename
        in: path
        required: true
        schema:
          type: string
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflows/character-master/{job_id}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Poll Character Master job
      operationId: api_character_master_get_api_v1_workflows_character_master__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflows/character-master/{job_id}/select:
    post:
      tags:
      - Control UI (legacy /api/v1)
      summary: Promote a Character Master candidate
      operationId: api_character_master_select_api_v1_workflows_character_master__job_id__select_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterMasterSelectRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assets/characters/{character_id}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Load persisted character asset
      operationId: api_character_asset_api_v1_assets_characters__character_id__get
      parameters:
      - name: character_id
        in: path
        required: true
        schema:
          type: string
          title: Character Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assets/characters/{character_id}/files/{filename}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Download character file
      operationId: api_character_asset_file_api_v1_assets_characters__character_id__files__filename__get
      parameters:
      - name: character_id
        in: path
        required: true
        schema:
          type: string
          title: Character Id
      - name: filename
        in: path
        required: true
        schema:
          type: string
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/generate/{job_id}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Poll generate job
      operationId: api_generate_get_api_v1_generate__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/upscale:
    post:
      tags:
      - Control UI (legacy /api/v1)
      summary: Queue RealESRGAN upscale
      operationId: api_upscale_post_api_v1_upscale_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpscaleRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/upscale/{job_id}/images/{filename}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Download upscale output image
      operationId: api_upscale_image_api_v1_upscale__job_id__images__filename__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: filename
        in: path
        required: true
        schema:
          type: string
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/upscale/{job_id}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Poll upscale job
      operationId: api_upscale_get_api_v1_upscale__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ltx-video:
    post:
      tags:
      - Control UI (legacy /api/v1)
      summary: Queue LTX-2.5 video
      description: t2v / i2v / a2v / flf2v / lipsync / motion_transfer on Comfy LTX :8189. Duration 1–30s.
        Camera is prompt-only. audioPrompt recommended.
      operationId: api_ltx_video_post_api_v1_ltx_video_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LtxVideoRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ltx-video/{job_id}/videos/{filename}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Download LTX MP4
      operationId: api_ltx_video_file_api_v1_ltx_video__job_id__videos__filename__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: filename
        in: path
        required: true
        schema:
          type: string
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ltx-video/{job_id}:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Poll LTX video job
      operationId: api_ltx_video_get_api_v1_ltx_video__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/downloads:
    get:
      tags:
      - Control UI (legacy /api/v1)
      summary: Catalog download status
      operationId: api_downloads_get_api_v1_downloads_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Api Downloads Get Api V1 Downloads Get
    post:
      tags:
      - Control UI (legacy /api/v1)
      summary: Start catalog download
      operationId: api_downloads_post_api_v1_downloads_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/profile:
    put:
      tags:
      - Control UI (legacy /api/v1)
      summary: Start exclusive GPU profile
      operationId: api_profile_put_api_v1_profile_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Control UI (legacy /api/v1)
      summary: Stop all GPU profiles
      operationId: api_profile_delete_api_v1_profile_delete
      responses:
        '204':
          description: Successful Response
...
components:
  schemas:
...
    CharacterMasterSelectRequest:
      properties:
        candidateIndex:
          type: integer
          maximum: 7.0
          minimum: 0.0
          title: Candidateindex
      type: object
      required:
      - candidateIndex
      title: CharacterMasterSelectRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UpscaleRequest:
      properties:
        image:
          anyOf:
          - type: string
          - type: 'null'
          title: Image
        sourceJobId:
          anyOf:
          - type: string
          - type: 'null'
          title: Sourcejobid
        sourceFilename:
          anyOf:
          - type: string
          - type: 'null'
          title: Sourcefilename
        scale:
          type: integer
          enum:
          - 2
          - 4
          title: Scale
          description: 'RealESRGAN scale factor: 2 or 4'
          default: 4
      type: object
      title: UpscaleRequest
...